+++ apps/tergite-mss/tests/api/rest/v2/test_calibrations.py
@@ -10,105 +10,234 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """Tests for calibrations v2"""
-import pytest
+from datetime import datetime, timedelta, timezone
+from typing import Any, Dict, List, Optional
 
-from tests._utils.date_time import get_current_timestamp_str
+import pytest
+from bson import ObjectId
+
+from tests._utils.date_time import get_current_timestamp_str, get_timestamp_str
 from tests._utils.fixtures import load_json_fixture
 from tests._utils.mongodb import find_in_collection, insert_in_collection
-from tests._utils.records import distinct_on, order_by, pop_field
+from tests._utils.records import (
+    distinct_on,
+    filter_by_equality,
+    order_by,
+    order_by_many,
+    with_current_timestamps,
+    with_incremental_timestamps,
+)
 
-_CALIBRATIONS_LIST = load_json_fixture("calibrations_v2.json")
+_CALIBRATIONS_LIST = load_json_fixture("calibrations.json")
 _LATEST_CALIBRATIONS = distinct_on(
     order_by(_CALIBRATIONS_LIST, field="last_calibrated", is_descending=True),
     field="name",
 )
 _DEVICE_NAMES = [item["name"] for item in _LATEST_CALIBRATIONS]
-_COLLECTION = "calibrations_v2"
+_COLLECTION = "calibrations"
 _LOGS_COLLECTION = "calibrations_logs"
 _EXCLUDED_FIELDS = ["_id"]
+_SKIP_LIMIT_SORT_PARAMS = [
+    (0, 1, ["-version", "last_calibrated"]),
+    (2, 4, None),
+    (1, 3, ["version"]),
+    (None, 10, None),
+    (2, None, ["last_calibrated"]),
+]
+_SEARCH_PARAMS = [
+    {"version": "2024.04.1"},
+    {"name": "Loke"},
+    {"version": "2023.06.0"},
+    {"version": "2023.06.0", "name": "Loke"},
+    {"version": "2023.06.0", "name": "Thor"},
+    {},
+]
+_PAGINATE_AND_SEARCH_PARAMS = [
+    (skip, limit, sort, search)
+    for skip, limit, sort in _SKIP_LIMIT_SORT_PARAMS
+    for search in _SEARCH_PARAMS
+]
 
 
-def test_read_calibrations(db, client_v2, user_jwt_cookie):
-    """GET `/v2/calibrations/` returns the latest calibrations"""
-    insert_in_collection(
-        database=db, collection_name=_COLLECTION, data=_CALIBRATIONS_LIST
+@pytest.mark.parametrize("skip, limit, sort, search", _PAGINATE_AND_SEARCH_PARAMS)
+def test_find_calibrations(
+    db,
+    client_v2,
+    skip: Optional[int],
+    limit: Optional[int],
+    sort: Optional[List[str]],
+    search: dict,
+    freezer,
+):
+    """Get to /v2/calibrations/?version=...&name=... can search for the calibrations that fulfill the given filters"""
+    raw_calibrations = with_incremental_timestamps(
+        _LATEST_CALIBRATIONS, fields=("last_calibrated", "updated_at")
     )
+    inserted_ids = insert_in_collection(
+        database=db, collection_name=_COLLECTION, data=raw_calibrations
+    )
+    raw_calibrations = _attach_str_ids(
+        raw_calibrations, ids=inserted_ids, id_field="id"
+    )
+
+    query_string = "?"
+    slice_end = len(raw_calibrations)
+    slice_start = 0
+    sort_fields = [
+        "-last_calibrated",
+    ]
+    if limit is not None:
+        query_string += f"limit={limit}&"
+        slice_end = limit
+    if skip is not None:
+        query_string += f"skip={skip}&"
+        slice_start = skip
+        slice_end += skip
+    if sort is not None:
+        sort_fields = sort
+        for sort_field in sort_fields:
+            query_string += f"sort={sort_field}&"
+
+    # Adding search
+    for key, value in search.items():
+        query_string += f"{key}={value}&"
 
     # using context manager to ensure on_startup runs
     with client_v2 as client:
-        response = client.get(f"/v2/calibrations", headers=user_jwt_cookie)
-        got = order_by(response.json(), field="name")
-        pop_field(got, "_id")
-        expected = order_by(_LATEST_CALIBRATIONS, field="name")
+        response = client.get(f"/v2/calibrations/{query_string}")
+        got = response.json()
+
+        filtered_data = filter_by_equality(raw_calibrations, filters=search)
+        sorted_data = order_by_many(filtered_data, fields=sort_fields)
+        expected = {
+            "skip": slice_start,
+            "limit": limit,
+            "data": sorted_data[slice_start:slice_end],
+        }
 
         assert response.status_code == 200
         assert got == expected
 
 
 @pytest.mark.parametrize("name", _DEVICE_NAMES)
-def test_read_calibration(name: str, db, client_v2, app_token_header):
+def test_read_calibration(name: str, db, client_v2, freezer):
     """Get `/v2/calibrations/{name}` reads the latest calibration of the given device"""
-    insert_in_collection(
-        database=db, collection_name=_COLLECTION, data=_CALIBRATIONS_LIST
+    raw_calibrations = with_incremental_timestamps(
+        _LATEST_CALIBRATIONS, fields=("last_calibrated", "updated_at")
+    )
+    inserted_ids = insert_in_collection(
+        database=db, collection_name=_COLLECTION, data=raw_calibrations
+    )
+    raw_calibrations = _attach_str_ids(
+        raw_calibrations, ids=inserted_ids, id_field="id"
     )
 
     # using context manager to ensure on_startup runs
     with client_v2 as client:
-        response = client.get(f"/v2/calibrations/{name}", headers=app_token_header)
-        got: dict = response.json()
-        got.pop("_id")
-        expected = list(filter(lambda x: x["name"] == name, _LATEST_CALIBRATIONS))[0]
+        response = client.get(f"/v2/calibrations/{name}")
+        got = response.json()
+        expected = filter_by_equality(raw_calibrations, {"name": name})[0]
 
         assert response.status_code == 200
-        assert expected == got
+        assert got == expected
 
 
-def test_create_calibrations(db, client_v2, system_app_token_header, freezer):
-    """POST calibrations to `/v2/calibrations` upserts them in calibrations_v2, and in calibrations_logs if not exist"""
+@pytest.mark.parametrize("payload", _CALIBRATIONS_LIST)
+def test_create(db, client_v2, system_app_token_header, payload, freezer):
+    """POST new calibration to `/v2/calibrations` creates it in calibrations and returns it"""
+    # using context manager to ensure on_startup runs
+    with client_v2 as client:
+        response = client.post(
+            "/v2/calibrations/",
+            json=payload,
+            headers=system_app_token_header,
+        )
+        got = response.json()
+        assert response.status_code == 200
+        assert got == {
+            **payload,
+            "updated_at": get_current_timestamp_str(),
+            "id": got["id"],
+        }
+
+
+@pytest.mark.parametrize("raw_payload", _CALIBRATIONS_LIST)
+def test_upsert(db, client_v2, system_app_token_header, raw_payload, freezer):
+    """POST an existing calibration to `/v2/calibrations` updates it in the calibrations collection"""
+    raw_calibrations = with_current_timestamps(
+        [raw_payload], fields=("updated_at", "last_calibrated")
+    )
+    insert_in_collection(
+        database=db, collection_name=_COLLECTION, data=raw_calibrations
+    )
+
     now = get_current_timestamp_str()
-    # insert only some of the calibrations in collection to show upsert is done
-    original_calibrations = _CALIBRATIONS_LIST[:2]
-    insert_in_collection(
-        database=db, collection_name=_COLLECTION, data=original_calibrations
+    future_timestamp = get_timestamp_str(
+        datetime.now(timezone.utc) + timedelta(hours=2)
     )
-    insert_in_collection(
-        database=db, collection_name=_LOGS_COLLECTION, data=original_calibrations
-    )
-    original_calibrations_in_db = find_in_collection(
-        db, collection_name=_COLLECTION, fields_to_exclude=_EXCLUDED_FIELDS
-    )
+    payload = {
+        **raw_payload,
+        "last_calibrated": future_timestamp,
+        "resonators": None,
+        "discriminators": None,
+    }
 
     # using context manager to ensure on_startup runs
     with client_v2 as client:
         response = client.post(
             "/v2/calibrations/",
-            json=_CALIBRATIONS_LIST,
+            json=payload,
             headers=system_app_token_header,
         )
-        final_calibrations_in_db = find_in_collection(
-            db, collection_name=_COLLECTION, fields_to_exclude=_EXCLUDED_FIELDS
+        got = response.json()
+        assert response.status_code == 200
+        assert got == {**payload, "updated_at": now, "id": got["id"]}
+
+        new_calibration_in_db = {
+            **payload,
+            "updated_at": now,
+            "_id": ObjectId(got["id"]),
+        }
+        final_db_data = find_in_collection(db, collection_name=_COLLECTION)
+        assert final_db_data == [new_calibration_in_db]
+
+
+@pytest.mark.parametrize("raw_payload", _CALIBRATIONS_LIST)
+def test_create_log(db, client_v2, system_app_token_header, raw_payload, freezer):
+    """POST calibrations to `/v2/calibrations` adds the calibration to calibrations_logs if not exist"""
+    raw_calibrations = with_current_timestamps(
+        _LATEST_CALIBRATIONS, fields=("updated_at",)
+    )
+    # insert only some of the calibrations in collection to show upsert is done
+    original_logs_in_db = raw_calibrations[:2]
+    insert_in_collection(
+        database=db, collection_name=_LOGS_COLLECTION, data=raw_calibrations[:2]
+    )
+
+    future_timestamp = get_timestamp_str(
+        datetime.now(timezone.utc) + timedelta(hours=2)
+    )
+    payload = {**raw_payload, "last_calibrated": future_timestamp}
+
+    # using context manager to ensure on_startup runs
+    with client_v2 as client:
+        response = client.post(
+            "/v2/calibrations/",
+            json=payload,
+            headers=system_app_token_header,
         )
-        final_calibration_logs_in_db = find_in_collection(
-            db, collection_name=_LOGS_COLLECTION, fields_to_exclude=_EXCLUDED_FIELDS
+        got = find_in_collection(
+            db, collection_name=_LOGS_COLLECTION, fields_to_exclude=("_id",)
         )
-        final_calibrations_in_db = order_by(final_calibrations_in_db, "name")
-        final_calibration_logs_in_db = order_by(final_calibration_logs_in_db, "name")
-        new_calibrations = [{**v, "last_calibrated": now} for v in _CALIBRATIONS_LIST]
-        expected_calibrations_in_db = order_by(new_calibrations, field="name")
-        expected_calibration_logs_in_db = order_by(
-            original_calibrations + new_calibrations[2:], field="name"
-        )
+        expected = [*original_logs_in_db, payload]
 
         assert response.status_code == 200
-        assert response.json() == "OK"
-
-        assert original_calibrations_in_db == original_calibrations
-        assert final_calibrations_in_db == expected_calibrations_in_db
-        assert final_calibration_logs_in_db == expected_calibration_logs_in_db
+        assert order_by(got, "name") == order_by(expected, "name")
 
 
-def test_create_calibrations_non_system_user(db, client_v2, user_jwt_cookie):
-    """Only system users can POST list of calibration-like dicts to `/v2/calibrations`"""
+@pytest.mark.parametrize("payload", _CALIBRATIONS_LIST)
+def test_create_calibrations_non_system_user(db, client_v2, payload, user_jwt_cookie):
+    """Only system users can POST calibration-like dicts to `/v2/calibrations`"""
     original_data_in_db = find_in_collection(
         db, collection_name=_COLLECTION, fields_to_exclude=_EXCLUDED_FIELDS
     )
@@ -117,7 +246,7 @@ def test_create_calibrations_non_system_user(db, client_v2, user_jwt_cookie):
     with client_v2 as client:
         response = client.post(
             "/v2/calibrations/",
-            json=_CALIBRATIONS_LIST,
+            json=payload,
             cookies=user_jwt_cookie,
         )
         final_data_in_db = find_in_collection(
@@ -130,3 +259,19 @@ def test_create_calibrations_non_system_user(db, client_v2, user_jwt_cookie):
 
         assert original_data_in_db == []
         assert final_data_in_db == []
+
+
+def _attach_str_ids(
+    data: List[Dict[str, Any]], ids: List[ObjectId], id_field: str = "_id"
+) -> List[Dict[str, Any]]:
+    """Attaches ids in string form to the data
+
+    Args:
+        data: the list of raw data records
+        ids: the list of ObjectId ids, in the same order as that of the data
+        id_field: the field on the records that is to have the id
+
+    Returns:
+        the list of records with ids attached to them
+    """
+    return [{**item, id_field: str(ids[idx])} for idx, item in enumerate(data)]